requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==3.0.1
dj-email-url==1.0.6
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.1
django-cache-url==3.4.5
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
environs==14.3.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
marshmallow==4.0.1
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
